requirements.txt
opencv-python==4.6.0.66
numpy
pillow==10.2.0